create mvvm and hit api

+++ app/src/main/java/com/example/weathermood/model/Current.kt
@@ -3,7 +3,7 @@ package com.example.weathermood.model
 data class Current(
     val clouds: Int,
     val dew_point: Double,
-    val dt: Int,
+    val dt: Long,
     val feels_like: Double,
     val humidity: Int,
     val pressure: Int,
@@ -15,5 +15,5 @@ data class Current(
     val visibility: Int,
     val weather: List<Weather>,
     val wind_deg: Int,
-    val wind_speed: Int
+    val wind_speed: Double
 )
